requirements.txt
Flask==2.0.1
Jinja2==3.0.1
python-dotenv==0.17.1
